noload: Do not attempt to compress tiny entries
Testing has shown that many small compression jobs may be sent to
noload_compress, this patch discards jobs < 64kB to prevent wasted
trafic to nvme-algo.  This may be a candidate for sw fallback if we want
tiny entry compression in the future and have a way to separtate tem on
decompression.