ZFS Offloading (ZOFF)
This commit contains the initial implementation of ZOFF. Hooks have been added into ZFS to allow for external, alternative implementations of compression, checksum, raidz, vdev_file, and vdev_disk in the write pipeline.

To use ZOFF
    1. Reconfigure ZFS with '--enable-zoff' and rebuild ZFS
    2. Load ZFS
    3. Build and load offloader

The vendor should only need to use the ZOFF provider API (sys/zoff.h). The functions in the zoff_functions struct should be implemented and passed into ZFS by the provider. The alternative implementations of operations are just that: alternative implementations, not alternative algorithms. This allows for the ZFS data to be accessible even if the provider is not.

The functions in the ZOFF provider API should be mostly stable. Major changes to them should not be necessary, but minor ones, such as renaming, reordering arguments, or reordering in the zoff_functions struct are possible.

An example provider name "example_zoff" can be found in the example_zoff directory. After building ZFS and loading the modules, run 'make' on the directory to build it. Load example_zoff.ko to enable it in ZFS. Note that example_zoff contains a large amount of ZFS code that was easier to copy than to export. The example_zoff also links with ZFS in order to not reimplement functionality that is exported and thus accessible. Additionally, due to the example offloader running in memory rather than on a different physical device, a lot of its functionality was conflated with or omittied from the provider's code, such as transfering function arguments and return values.

TODO: Create a richer set of error codes, and update the existing returns.
TODO: Remove modifications to ZFS that are only needed by the example kernel offloader.